tests, tests

+++ ch.opentrainingcenter.client/src/ch/opentrainingcenter/client/model/ModelFactory.java
@@ -18,18 +18,18 @@ public final class ModelFactory {
      * Erstellt ein SimpleTraining mit dem Lauf Typ NONE
      */
     public static ISimpleTraining createSimpleTraining(final ITraining overview) {
-        return new SimpleTraining(overview.getLaengeInMeter(), overview.getDauerInSekunden(), overview.getDateOfStart(), overview.getAverageHeartBeat(),
-                overview.getMaxHeartBeat(), overview.getMaxSpeed(), RunType.NONE);
+        return new SimpleTraining(overview.getLaengeInMeter(), overview.getDauerInSekunden(), overview.getDateOfStart(), overview
+                .getAverageHeartBeat(), overview.getMaxHeartBeat(), overview.getMaxSpeed(), RunType.NONE);
     }
 
-    public static ISimpleTraining createSimpleTraining(final double distanzInMeter, final double dauerInSekunden, final Date datum, final int avgHeartRate, final int maxHeartRate,
-            final int maxSpeed, final RunType type) {
+    public static ISimpleTraining createSimpleTraining(final double distanzInMeter, final double dauerInSekunden, final Date datum,
+            final int avgHeartRate, final int maxHeartRate, final double maxSpeed, final RunType type) {
         return new SimpleTraining(distanzInMeter, dauerInSekunden, datum, avgHeartRate, maxHeartRate, maxSpeed, type);
     }
 
     public static ISimpleTraining createSimpleTraining(final ITraining overview, final RunType runType) {
-        return new SimpleTraining(overview.getLaengeInMeter(), overview.getDauerInSekunden(), overview.getDateOfStart(), overview.getAverageHeartBeat(),
-                overview.getMaxHeartBeat(), overview.getMaxSpeed(), runType);
+        return new SimpleTraining(overview.getLaengeInMeter(), overview.getDauerInSekunden(), overview.getDateOfStart(), overview
+                .getAverageHeartBeat(), overview.getMaxHeartBeat(), overview.getMaxSpeed(), runType);
     }
 
     public static IGpsFileModel createGpsFileModel(final String fileName) {
